+++ .gitignore
@@ -1,1 +1,2 @@
 src/main/resources/application.properties
+.idea
