+++ pom.xml
@@ -54,8 +54,9 @@
         <dependency>
             <groupId>org.postgresql</groupId>
             <artifactId>postgresql</artifactId>
-            <scope>runtime</scope>
+            <version>42.7.3</version>
         </dependency>
+
         <!--Alexandr Bezarov H2 DB -->
         <dependency>
             <groupId>com.h2database</groupId>
