--- src/main/java/com/coolbank/dto/AccountDto.java
@@ -1,27 +1,0 @@
-package com.coolbank.dto;
-
-public class AccountDto {
-    private String accountNumber;
-    private Double balance;
-
-    public AccountDto(String accountNumber, Double balance) {
-        this.accountNumber = accountNumber;
-        this.balance = balance;
-    }
-
-    public String getAccountNumber() {
-        return accountNumber;
-    }
-
-    public void setAccountNumber(String accountNumber) {
-        this.accountNumber = accountNumber;
-    }
-
-    public Double getBalance() {
-        return balance;
-    }
-
-    public void setBalance(Double balance) {
-        this.balance = balance;
-    }
-}
