+++ src/main/java/com/coolbank/dto/PaymentDTO.java
@@ -1,17 +1,18 @@
 package com.coolbank.dto;
 
 import com.coolbank.model.Account;
+import com.coolbank.model.Payment;
 
 import java.time.LocalDateTime;
 
-public class PaymentDto {
+public class PaymentDTO {
     private Double amount;
     private Account fromAccount;
     private Account toAccount;
     private LocalDateTime paymentDate;
     private String status;
 
-    public PaymentDto(Double amount, Account fromAccount, Account toAccount, LocalDateTime paymentDate, String status) {
+    public PaymentDTO(Double amount, Account fromAccount, Account toAccount, LocalDateTime paymentDate, String status) {
         this.amount = amount;
         this.fromAccount = fromAccount;
         this.toAccount = toAccount;
@@ -57,5 +58,11 @@ public class PaymentDto {
 
     public void setStatus(String status) {
         this.status = status;
+    }
+
+    public static PaymentDTO mappToPaymentDTO(Payment payment){
+        return new PaymentDTO(payment.getAmount(),
+                payment.getFromAccount(), payment.getToAccount(),
+                payment.getPaymentDate(), payment.getStatus());
     }
 }
