+++ src/main/resources/application.properties
@@ -39,3 +39,6 @@ src/main/java/com/coolbank/Log/*
 src/main/resources/*
 src/main/resources/application.properties
 src/main/resources/data.sql
+##############################
+## PostgreSQL DB
+##############################
